Basic Win/Lose conditions and related updates
Players can now pick up objects tagged as "Key", which will implement the key count in the UI. As long as there is at least one key in the key count, players can pass through (destroy) objects tagged as "LockedDoor". If both players reach zero health, a game over screen is brought up with "Give Up" and "Try Again" buttons that quit the game or reload the scene respectively. Alternatively, if the player touches an object tagged as "Treasure", it brings up a Game Win screen instead, with buttons labeled "Finish" and "Play Again" which reuse the same functionality as the lose buttons.